third_party/WebKit/LayoutTests/fast/forms/select-popup/popup-menu-size-expected.txt

PASS popupWindow.innerWidth is 100
PASS successfullyParsed is true

TEST COMPLETE
